WIP: web: extract component for new VG to a separate file
